rewrite it kinda :sob: sorry pufferfish you did the hard work tho also i didnt really rewrite that much it was mainly orginization and stuff
