+++ common/games.js
@@ -1,13 +1,17 @@
 import { Collection, ButtonBuilder, ButtonStyle } from "discord.js";
-import { MessageActionRowBuilder } from "../../types/ActionRowBuilder.js";
+import { MessageActionRowBuilder } from "../types/ActionRowBuilder.js";
 
-/** @type {Collection<string, import("discord.js").Message>} */
+/** @type {Collection<string, string>} */
 export const CURRENTLY_PLAYING = new Collection();
 
 /**
  * Reply to the interaction if the interaction user is already playing a game.
  *
- * @param {import("discord.js").MessageComponentInteraction | import("discord.js").CommandInteraction} interaction - The interaction to analyze.
+ * @param {| import("discord.js").CommandInteraction
+ * 	| import("discord.js").MessageComponentInteraction
+ * 	| import("discord.js").ModalSubmitInteraction} interaction
+ *   - The interaction to analyze.
+ *
  *
  * @returns {Promise<boolean>} Whether or not the user is already playing.
  */
@@ -22,11 +26,7 @@ export async function checkIfUserPlaying(interaction) {
 				new ButtonBuilder()
 					.setLabel("Go to game")
 					.setStyle(ButtonStyle.Link)
-					.setURL(
-						`https://discord.com/channels/${encodeURI(
-							current.guild?.id || "@me",
-						)}/${encodeURI(current.channel.id)}/${encodeURI(current.id)}`,
-					),
+					.setURL(current),
 			),
 		],
