+++ lib/text.js
@@ -85,3 +85,14 @@ export function normalize(text) {
 			"",
 		);
 }
+
+/**
+ * Trims the patch version off of a Semver.
+ *
+ * @param {string} full - The full version.
+ *
+ * @returns {string} - The patchless version.
+ */
+export function trimPatchVersion(full) {
+	return /^(?<main>\d+\.\d+)\.\d+/.exec(full)?.groups?.main || "";
+}
